Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a, b, c;
-    scanf("%d %d %d", &a, &b, &c);
-    int sum=a+b+c;
+    float a, b, c;
+    scanf("%f %f %f", &a, &b, &c);
+    float sum=a+b+c;
     float average=sum/3;
     printf("Average: %.2f\n",average);
     return 0;
